test-requirements.txt
openstacksdk
pytest
shortuuid
six
